requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra-index-url=https://download.pytorch.org/whl/cpu --strip-extras requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cpu

annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   openai
async-timeout==4.0.3
    # via redis
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
distro==1.9.0
    # via openai
einops==0.8.0
    # via -r requirements.in
exceptiongroup==1.2.2
    # via anyio
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2024.9.0
    # via
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via
    #   langsmith
    #   openai
huggingface-hub==0.25.2
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via torch
jiter==0.6.1
    # via openai
joblib==1.4.2
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.3.10
    # via langchain-text-splitters
langchain-text-splitters==0.3.0
    # via -r requirements.in
langsmith==0.1.134
    # via langchain-core
markupsafe==3.0.1
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.1
    # via torch
numpy==2.1.2
    # via
    #   -r requirements.in
    #   scikit-learn
    #   scipy
    #   transformers
openai==1.51.2
    # via -r requirements.in
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   huggingface-hub
    #   langchain-core
    #   transformers
pillow==10.4.0
    # via sentence-transformers
pydantic==2.9.2
    # via
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.23.4
    # via pydantic
pymupdf==1.24.11
    # via pymupdf4llm
pymupdf4llm==0.0.17
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain-core
    #   transformers
redis==5.1.1
    # via -r requirements.in
regex==2024.9.11
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   langsmith
    #   requests-toolbelt
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
safetensors==0.4.5
    # via transformers
scikit-learn==1.5.2
    # via sentence-transformers
scipy==1.14.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.2.0
    # via -r requirements.in
slack-sdk==3.33.1
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sympy==1.13.3
    # via torch
tenacity==8.5.0
    # via langchain-core
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.20.1
    # via transformers
torch==2.4.1+cpu
    # via
    #   -r requirements.in
    #   sentence-transformers
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.45.2
    # via
    #   -r requirements.in
    #   sentence-transformers
typing-extensions==4.12.2
    # via
    #   anyio
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.2.3
    # via requests
